+++ commons/src/main/java/org/parsingbot/service/impl/VacancyServiceImpl.java
@@ -31,7 +31,7 @@ public class VacancyServiceImpl implements VacancyService {
     }
 
     @Override
-    public List<Vacancy> getVacanciesByUserId(Integer userId) {
+    public List<Vacancy> getVacanciesByUserId(Long userId) {
         List<Integer> userVacanciesIds = userRepository.findUserVacanciesIds(userId);
         return getVacanciesByIds(userVacanciesIds);
     }
